requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
cerebras_cloud_sdk==1.6.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
distro==1.9.0
duckduckgo_search==6.3.1
EditorConfig==0.12.4
entrypoints==0.4
executing==2.1.0
Faker==30.6.0
favicon==0.7.0
fonttools==4.54.1
fpdf==1.7.2
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.41
groq==0.11.0
h11==0.14.0
htbuilder==0.6.2
html2text==2024.2.26
httpcore==1.0.6
httpx==0.27.2
idna==3.10
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.6.1
jsbeautifier==1.15.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==3.0.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.5.0
multidict==6.1.0
narwhals==1.9.3
nest-asyncio==1.6.0
numpy==2.1.2
openai==1.51.2
packaging==24.1
pandas==2.2.3
parso==0.8.4
pdfkit==1.0.0
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
primp==0.6.4
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.2
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pymdown-extensions==10.11.2
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rich==13.9.2
rpds-py==0.20.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
st-annotated-text==4.0.1
st-theme==1.2.3
stack-data==0.6.3
streamlit==1.39.0
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.5.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-image-gallery==0.0.2
streamlit-keyup==0.2.4
streamlit-shadcn-ui==0.1.18
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tenacity==9.0.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
validators==0.34.0
watchdog==5.0.3
wcwidth==0.2.13
wkhtmltopdf==0.2
yarl==1.15.4
